requirements.txt
alembic==1.0.11
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
Click==7.0
coverage==4.5.3
entrypoints==0.3
filelock==3.0.12
flake8==3.7.7
Flask==1.0.3
Flask-Cors==3.0.8
Flask-GraphQL==2.0.0
Flask-JSON==0.3.3
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
graphene==2.1.6
graphene-sqlalchemy==2.2.1
graphql-core==2.2
graphql-relay==0.4.5
graphql-server-core==1.1.1
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==19.0
pluggy==0.12.0
promise==2.2.1
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
gunicorn==19.9.0
PyJWT==1.6.4
pyparsing==2.4.0
pytest==4.6.3
pytest-cov==2.7.1
python-dateutil==2.8.0
python-editor==1.0.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.5
SQLAlchemy-Utils==0.34.0
toml==0.10.0
tox==3.13.1
validators==0.12.4
virtualenv==16.6.1
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
